+++ RELEASES.md
@@ -14,6 +14,7 @@ Upcoming
 
 - Add support for integer and boolean parameters when using QuerySenders
 - Fix broken "Show configuration warnings only at relevant configuration"
+- Show config name for configuration errors and warnings